Fix bugs migration

+++ database/migrations/create_simgtk_sekolah_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create(config('simgtk.table_prefix').'sekolah', function (Blueprint $table) {
+        Schema::create(config('simgtk.table_prefix') . 'sekolah', function (Blueprint $table) {
             $table->ulid('id')->primary();
             $table->string('nama');
             $table->string('npsn')->index();
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('wilayah_kode')->nullable()->index();
             $table->unsignedTinyInteger('jumlah_kelas')->nullable();
             $table->unsignedTinyInteger('jumlah_rombel')->nullable();
-            $table->unsignedTinyInteger('jumlah_siswa')->nullable();
+            $table->unsignedSmallInteger('jumlah_siswa')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
@@ -30,6 +30,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists(config('simgtk.table_prefix').'sekolah');
+        Schema::dropIfExists(config('simgtk.table_prefix') . 'sekolah');
     }
 };
